feat: refactoring manual balances store
