+++ frontend/app/src/services/monitoring.ts
@@ -3,6 +3,7 @@ import { storeToRefs } from 'pinia';
 import { setupGeneralBalances } from '@/composables/balances';
 import { websocket } from '@/services/websocket/websocket-service';
 import { useExchangeBalancesStore } from '@/store/balances/exchanges';
+import { useManualBalancesStore } from '@/store/balances/manual';
 import { useNotifications } from '@/store/notifications';
 import { useSessionStore } from '@/store/session';
 import { useWatchersStore } from '@/store/session/watchers';
@@ -30,12 +31,10 @@ class Monitoring {
   }
 
   private static async fetchBalances() {
-    const {
-      fetchManualBalances,
-      fetchBlockchainBalances,
-      fetchLoopringBalances,
-      refreshPrices
-    } = setupGeneralBalances();
+    const { fetchBlockchainBalances, fetchLoopringBalances, refreshPrices } =
+      setupGeneralBalances();
+
+    const { fetchManualBalances } = useManualBalancesStore();
     const { fetchConnectedExchangeBalances } = useExchangeBalancesStore();
 
     await fetchManualBalances();
