+++ frontend/common/package.json
@@ -37,7 +37,7 @@
     "npm": ">=8 <9"
   },
   "dependencies": {
-    "bignumber.js": "9.0.2",
+    "bignumber.js": "9.1.0",
     "zod": "3.12.0"
   },
   "peerDependencies": {
